v0.1
Пофикшен баг с неверным количеством участников при старте бота
Пофикшен баг с неверным отображением участников в userinfo
Пофикшен баг с тем, что в userinfo не было иконки в эмбеде
Создана команда со списком всех серверов, где находится бот
Создана команда для лива с определенного сервера
Пофикшена проблема с иконками в help
В userinfo добавлена поддержка поиска по нику/айди
Добавлена поддержка спец. символов в нике в userinfo
Код стал более читаемый
Исправлена работа команд allnicks и resetallnicks: не работают на ботов, владельца и стоящих выше по иерархии ролей
В avatar исправлен поиск по нику/айди
Исправлена работа clear

+++ commands/admin/say.js
@@ -1,5 +1,9 @@
 const Discord = require("discord.js");
-module.exports =  async function(message,args){ if (!message.channel.permissionsFor(message.member).has("ADMINISTRATOR", false)) return message.reply('вы не администратор.');
-const sayMessage = args.join(" ");
-message.delete().catch(O_o => { });
-message.channel.send(sayMessage);}
+module.exports =  async function(message,args)
+{
+    if (!message.channel.permissionsFor(message.member).has("ADMINISTRATOR", false))
+        return message.reply('Вы не администратор.');
+    const sayMessage = args.join(" ");
+    message.delete().catch(O_o => { });
+    message.channel.send(sayMessage);
+}
